+++ packages/router/src/handler.ts
@@ -41,6 +41,10 @@ import { Subgraph } from "./subgraph";
 
 export const EXPIRY_DECREMENT = 3600 * 24;
 export const SWAP_RATE = "0.9995"; // 0.05% fee
+export const ONE_DAY_IN_SECONDS = 3600 * 24;
+
+/** Determine if expiry is valid */
+export const validExpiry = (expiry: number) => expiry - Math.floor(Date.now() / 1000) > ONE_DAY_IN_SECONDS;
 
 export interface TransactionDataParams {
   user: string;
@@ -82,6 +86,7 @@ export class HandlerError extends NxtpError {
       methodId: string;
       method: string;
       calling: string;
+      cancellable?: boolean;
     },
   ) {
     super(message, context, HandlerError.type);
@@ -123,9 +128,6 @@ export const mutateAmount = (amount: string) => {
  */
 export const mutateExpiry = (expiry: number): number => {
   const rxExpiry = expiry - EXPIRY_DECREMENT;
-  if (rxExpiry < Date.now() / 1000) {
-    throw new Error("Expiration already happened");
-  }
   return rxExpiry;
 };
 
@@ -144,7 +146,7 @@ export class Handler {
   private senderFulfilling: Map<string, boolean> = new Map();
   constructor(
     private readonly messagingService: RouterNxtpNatsMessagingService,
-    private readonly _subgraph: Subgraph,
+    private readonly subgraph: Subgraph,
     private readonly txManager: TransactionManager,
     private readonly txService: TransactionService,
     private readonly signer: Wallet,
@@ -230,7 +232,7 @@ export class Handler {
     const sendingConfig = config.chainConfig[sendingChainId];
     const receivingConfig = config.chainConfig[receivingChainId];
     let bid: AuctionBid;
-    const result = await this._subgraph
+    const result = await this.subgraph
       .getAssetBalance(receivingAssetId, receivingChainId)
       .andThen((balance) => {
         // validate liquidity
@@ -564,6 +566,7 @@ export class Handler {
     // that user is only sending stuff that makes sense is possibly ok since otherwise
     // they're losing gas costs
 
+    const receiverExpiry = mutateExpiry(txData.expiry);
     let bid: AuctionBid;
     const validationRes = Result.fromThrowable(
       decodeAuctionBid,
@@ -591,7 +594,7 @@ export class Handler {
             }),
         )(bid, bidSignature);
       })
-      .andThen((recovered) => {
+      .andThen<undefined, HandlerError>((recovered) => {
         if (recovered !== this.signer.address) {
           return err(
             new HandlerError(HandlerError.reasons.PrepareValidationError, {
@@ -605,7 +608,7 @@ export class Handler {
         }
         return ok(undefined);
       })
-      .andThen(() => {
+      .andThen<undefined, HandlerError>(() => {
         // TODO: anything else? seems unnecessary to validate everything
         if (!BigNumber.from(bid.amount).eq(txData.amount) || bid.transactionId !== txData.transactionId) {
           return err(
@@ -620,19 +623,21 @@ export class Handler {
         }
         return ok(undefined);
       })
-      .andThen(() =>
-        Result.fromThrowable(
-          mutateExpiry,
-          (err) =>
+      .andThen<undefined, HandlerError>(() => {
+        if (!validExpiry(receiverExpiry)) {
+          return err(
             new HandlerError(HandlerError.reasons.PrepareValidationError, {
               method,
               methodId,
-              calling: "mutateExpiry",
+              calling: "",
               requestContext,
-              prepareError: (err as Error).message,
+              prepareError: "receiverExpiry < txData.expiry",
+              cancellable: true,
             }),
-        )(txData.expiry),
-      );
+          );
+        }
+        return ok(undefined);
+      });
 
     if (validationRes.isOk()) {
       this.logger.info({ method, methodId, requestContext }, "Validated input");
@@ -647,6 +652,28 @@ export class Handler {
         },
         "Error during validation",
       );
+
+      if (validationRes.error.context.cancellable) {
+        this.logger.info({ method, methodId, requestContext }, "Cancelling transaction");
+        const cancelRes = await this.txManager.cancel(
+          txData.sendingChainId,
+          {
+            relayerFee: "0",
+            txData,
+            signature: "0x",
+          },
+          requestContext,
+        );
+        if (cancelRes.isOk()) {
+          this.logger.info({ method, methodId, requestContext }, "Cancelled transaction");
+        } else {
+          this.logger.error(
+            { method, methodId, requestContext, err: jsonifyError(cancelRes.error) },
+            "Error cancelling transaction",
+          );
+        }
+      }
+      this.receiverPreparing.delete(txData.transactionId);
       return;
     }
 
@@ -681,7 +708,7 @@ export class Handler {
       {
         txData,
         amount: mutateAmount(txData.amount),
-        expiry: mutateExpiry(txData.expiry),
+        expiry: receiverExpiry,
         bidSignature,
         encodedBid,
         encryptedCallData,