+++ packages/utils/src/index.ts
@@ -14,3 +14,4 @@ export * from "./hexStrings";
 export * from "./transactionManager";
 export * from "./request";
 export * from "./mock";
+export * from "./subgraph";