+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-# NXTP Changelog
+# Vector Changelog
 
 ## Next Release
+
+## 0.0.16
+
+- [sdk] Fix cancellation.
+- [sdk] Change method name from `cancelExpired` to `cancel`.
+- [sdk, router] Add `network` param for abstracting messaging configuration.
+- [sdk] Rename `startTransfer` and `finishTransfer` to `prepareTransfer` and `fulfillTransfer`.
 
 ## 0.0.15